Make private variables more explicit.
The goal of this commit is to private clarity to the reader. The only implicitly defined remaining are the static constructors for UnityEditorShellWindow.Styles and Autocomplete.Styles, because it's against the (Unity?) C# compiler.

Add spaces after 'if'

Move if brackets to own line, make all if's have brackets.

Fix misc code inconsistencies